Float16 (#310)
* Fix casting to use the default floatx where possible to avoid type
errors when training using mixed precision or float16.

* Update tests for supporting float16

* Remove float dtype parameterization of readme tests. They were too slow.

* Fix casting error when passing constant scalar. Set policy in multihead test to ensure we reset the policy to a good state.

* Remove duplicate long running test. This should speed up test by ~3min.

+++ tensorflow_similarity/losses/barlow.py
@@ -80,7 +80,7 @@ class Barlow(tf.keras.losses.Loss):
 
         # compute pairwise
         c = tf.matmul(za, zb, transpose_a=True)
-        c = c / tf.cast(batch_size, dtype="float32")
+        c = c / tf.cast(batch_size, dtype=c.dtype)
 
         on_diag = 1.0 - tf.linalg.diag_part(c)
         on_diag = tf.math.pow(on_diag, 2)
